third_party/blink/web_tests/external/wpt/html/semantics/selectors/pseudo-classes/disabled-expected.txt

This is a testharness.js-based test.
FAIL ':disabled' should match only disabled elements assert_array_equals: lengths differ, expected array [Element node <button id="button2" disabled="">button2</button>, Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <fieldset disabled="" id="fieldset2">
  <legend><input ty..., Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 9, got [Element node <button id="button2" disabled="">button2</button>, Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 8
FAIL ':disabled' should not match elements whose disabled attribute has been removed assert_array_equals: lengths differ, expected array [Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <fieldset disabled="" id="fieldset2">
  <legend><input ty..., Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 8, got [Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 7
FAIL ':disabled' should also match elements whose disabled attribute has been set assert_array_equals: lengths differ, expected array [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <fieldset disabled="" id="fieldset2">
  <legend><input ty..., Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 9, got [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 8
FAIL ':disabled' should also match elements whose disabled attribute has been set twice assert_array_equals: lengths differ, expected array [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <fieldset disabled="" id="fieldset2">
  <legend><input ty..., Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 9, got [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled=""></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 8
FAIL ':disabled' should also match disabled elements whose type has changed assert_array_equals: lengths differ, expected array [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled="" type="submit"></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <fieldset disabled="" id="fieldset2">
  <legend><input ty..., Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 9, got [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled="" type="submit"></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 8
FAIL ':disabled' should not match elements not in the document assert_array_equals: lengths differ, expected array [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled="" type="submit"></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <fieldset disabled="" id="fieldset2">
  <legend><input ty..., Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 9, got [Element node <button id="button1" type="submit" disabled="disabled">bu..., Element node <input id="input2" disabled="" type="submit"></input>, Element node <select disabled="" id="select2">
 <optgroup label="optio..., Element node <optgroup label="options" disabled="" id="optgroup2">
  <..., Element node <option value="option2" disabled="" id="option2">option2
..., Element node <textarea disabled="" id="textarea2">textarea2</textarea>, Element node <input id="clubname" required=""></input>, Element node <input id="clubnum" required="" pattern="[-0-9]+"></input>] length 8
Harness: the test ran to completion.
